kirana_kart/requirements.txt
# ============================================================
# Kirana Kart — Python Dependencies
# ============================================================

# ----------------------------
# Web Framework
# ----------------------------
fastapi>=0.111.0,<0.120.0
uvicorn[standard]>=0.29.0,<0.35.0
python-multipart>=0.0.9        # file upload support

# ----------------------------
# Database
# ----------------------------
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0

# ----------------------------
# Configuration
# ----------------------------
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# ----------------------------
# Redis + Task Queue
# ----------------------------
redis>=5.0.0,<6.0.0
celery[redis]>=5.4.0,<6.0.0

# ----------------------------
# LLM / AI
# ----------------------------
openai>=1.30.0,<2.0.0

# ----------------------------
# Vector Database
# ----------------------------
weaviate-client>=3.26.7,<4.0.0

# ----------------------------
# Document Processing
# ----------------------------
python-docx>=1.1.0,<2.0.0
markdown>=3.6,<4.0

# ----------------------------
# Logging (structured / JSON)
# ----------------------------
python-json-logger>=2.0.7,<3.0.0

# ----------------------------
# Observability — Prometheus
# ----------------------------
prometheus-client>=0.20.0,<1.0.0

# ----------------------------
# Observability — OpenTelemetry
# ----------------------------
opentelemetry-sdk>=1.25.0,<2.0.0
opentelemetry-api>=1.25.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-sqlalchemy>=0.46b0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0,<2.0.0

# ----------------------------
# Testing
# ----------------------------
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0           # async-compatible TestClient for FastAPI
